modulo-recetas/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
jinja2==3.1.2
python-dotenv==1.0.0
httpx==0.23.3
supabase==1.0.3
openai==1.3.0
python-multipart==0.0.6
pydantic==2.4.2